renamed component_models.py to base_component.py; added not implemented methods to component models for more clarification
