+++ pandapipes/component_models/abstract_models/__init__.py
@@ -1,4 +1,4 @@
-from .component_models import *
+from .base_component import *
 from .branch_models import *
 from .branch_w_internals_models import *
 from .branch_wo_internals_models import *
