+++ pandapipes/component_models/abstract_models/branch_models.py
@@ -29,6 +29,18 @@ logger = logging.getLogger(__name__)
 class BranchComponent(Component):
 
     @classmethod
+    def table_name(cls):
+        raise NotImplementedError
+
+    @classmethod
+    def get_component_input(cls):
+        raise NotImplementedError
+
+    @classmethod
+    def get_result_table(cls, net):
+        raise NotImplementedError
+
+    @classmethod
     def active_identifier(cls):
         raise NotImplementedError()
