+++ pandapipes/component_models/abstract_models/branch_w_internals_models.py
@@ -23,6 +23,18 @@ class BranchWInternalsComponent(BranchComponent):
     """
 
     @classmethod
+    def active_identifier(cls):
+        raise NotImplementedError
+
+    @classmethod
+    def calculate_pressure_lift(cls, net, pipe_pit, node_pit):
+        raise NotImplementedError
+
+    @classmethod
+    def calculate_temperature_lift(cls, net, pipe_pit, node_pit):
+        raise NotImplementedError
+
+    @classmethod
     def internal_node_name(cls):
         return NotImplementedError
