+++ pandapipes/component_models/abstract_models/branch_wzerolength_models.py
@@ -20,6 +20,30 @@ class BranchWZeroLengthComponent(BranchWOInternalsComponent):
     """
 
     @classmethod
+    def table_name(cls):
+        raise NotImplementedError
+
+    @classmethod
+    def get_component_input(cls):
+        raise NotImplementedError
+
+    @classmethod
+    def get_result_table(cls, net):
+        raise NotImplementedError
+
+    @classmethod
+    def active_identifier(cls):
+        raise NotImplementedError
+
+    @classmethod
+    def calculate_pressure_lift(cls, net, pipe_pit, node_pit):
+        raise NotImplementedError
+
+    @classmethod
+    def calculate_temperature_lift(cls, net, pipe_pit, node_pit):
+        raise NotImplementedError
+
+    @classmethod
     def create_pit_branch_entries(cls, net, branch_wzerolength_pit, node_name):
         """
         Function which creates pit branch entries with a specific table.
