+++ pandapipes/component_models/abstract_models/node_models.py
@@ -2,7 +2,7 @@
 # and Energy System Technology (IEE), Kassel. All rights reserved.
 # Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
 
-from pandapipes.component_models.abstract_models.component_models import Component
+from pandapipes.component_models.abstract_models.base_component import Component
 
 try:
     import pplog as logging
